Building the Home Page, including Navbar, Footer, and Header
Having the index page done now it created a good understanding for how the flow of the rest of the pages are going to go. As the Header, Navbar and Footer will continue to be the exact same across all pages.  This uniform perspective has allowed me to see what works and doesn't work, and many changes can be seen on the edited code.